🚚refactor: directory structure
